Move Plated instances to monocle.std
Pulled in the Cofree addition to prevent merge conflicts later on.
