+++ core/src/main/scala/monocle/std/Cofree.scala
@@ -1,7 +1,7 @@
 package monocle.std
 
 import monocle.function._
-import monocle.{Lens, PIso, Iso, Optional, PTraversal, Traversal}
+import monocle.{Lens, PIso, Iso, Optional, PTraversal, Plated, Traversal}
 
 import scalaz.Cofree._
 import scalaz.{Cofree, Free, Applicative, Traverse, OneAnd, Tree}
@@ -63,4 +63,12 @@ trait CofreeOptics {
   /** Trivial [[Each]] instance due to `Cofree S` being traversable when `S` is */
   implicit def cofreeEach[S[_] : Traverse, A]: Each[Cofree[S, A], A] =
     Each.traverseEach[({type L[X] = Cofree[S, X]})#L, A]
+
+  implicit def cofreePlated[S[_]: Traverse, A]: Plated[Cofree[S, A]] = new Plated[Cofree[S, A]] {
+    def plate: Traversal[Cofree[S, A], Cofree[S, A]] = new Traversal[Cofree[S, A], Cofree[S, A]] {
+      def modifyF[F[_]: Applicative](f: Cofree[S, A] => F[Cofree[S, A]])(s: Cofree[S, A]): F[Cofree[S, A]] =
+        Applicative[F].map(Traverse[S].traverse(s.t.run)(f))(Cofree(s.head, _))
+    }
+  }
+
 }
