+++ core/src/main/scala/monocle/std/IList.scala
@@ -1,7 +1,7 @@
 package monocle.std
 
 import monocle.function._
-import monocle.{Iso, Optional, PIso, Prism}
+import monocle.{Iso, Optional, PIso, Plated, Prism, Traversal}
 
 import scalaz.Id.Id
 import scalaz.std.option._
@@ -57,5 +57,15 @@ trait IListInstances {
 
   implicit def iListReverse[A]: Reverse[IList[A], IList[A]] =
     Reverse.reverseFromReverseFunction[IList[A]](_.reverse)
+
+  implicit def ilistPlated[A]: Plated[IList[A]] = new Plated[IList[A]] {
+    def plate: Traversal[IList[A], IList[A]] = new Traversal[IList[A], IList[A]] {
+      def modifyF[F[_]: Applicative](f: IList[A] => F[IList[A]])(s: IList[A]): F[IList[A]] =
+        s match {
+          case ICons(x, xs) => Applicative[F].map(f(xs))(x :: _)
+          case INil() => Applicative[F].point(INil())
+        }
+    }
+  }
 
 }
